refactor: Update Zendesk integration in react-thirdparty package
