+++ packages/react-thirdparty/recaptcha/index.ts
@@ -1,5 +1,4 @@
 export * from "./useGoogleReCaptcha";
-export * from "./useGoogleReCaptchaContext";
 export * from "./GoogleReCaptcha";
 export * from "./Context";
 export * from "./Provider";
