--- packages/react-thirdparty/recaptcha/useGoogleReCaptchaContext.ts
@@ -1,5 +1,0 @@
-import { useContext } from "react";
-import { GoogleReCaptchaContext } from "./Context";
-
-export const useGoogleReCaptchaContext = () =>
-  useContext(GoogleReCaptchaContext);
