+++ packages/react-thirdparty/zendesk/Context.tsx
@@ -1,22 +1,14 @@
 import { createContext, useContext } from "react";
 
 export interface ZendeskContextProps {
-  executeZendesk: ZendeskWidget;
+  executeZendesk?: ZendeskWidget;
 }
 
-export const ZendeskContext = createContext<ZendeskContextProps | undefined>(
-  undefined
-);
+export const ZendeskContext = createContext<ZendeskContextProps>({});
 
 /**
  * Custom hook to use the Zendesk context.
  * @throws If used outside of a ZendeskProvider.
  * @returns The Zendesk context value.
  */
-export const useZendeskContext = (): ZendeskContextProps => {
-  const context = useContext(ZendeskContext);
-  if (!context) {
-    throw new Error("useZendeskContext must be used within a ZendeskProvider");
-  }
-  return context;
-};
+export const useZendeskContext = () => useContext(ZendeskContext);
