+++ packages/react-thirdparty/zendesk/useZendesk.ts
@@ -4,7 +4,4 @@ import { useZendeskContext } from "./Context";
  * Custom hook to access Zendesk client through context.
  * @returns The Zendesk context value.
  */
-export const useZendesk = () => {
-  const context = useZendeskContext();
-  return context;
-};
+export const useZendesk = () => useZendeskContext();
